requirements.txt
# Movie Recommendation System Requirements
# Python 3.10+

# Install instructions:
# python -m venv venv
# source venv/bin/activate      (Linux/Mac)
# venv\Scripts\activate         (Windows)
# pip install -r requirements.txt

# Core libraries
pandas>=2.0,<2.3
numpy>=1.24,<2.0

# Visualization
matplotlib>=3.7,<3.9
seaborn>=0.12,<0.14

# Machine Learning
scikit-learn>=1.3,<1.6
scipy>=1.10,<1.13

# Recommender Systems
scikit-surprise>=1.1.3,<1.2

# Notebook
jupyter>=1.0,<2.0
notebook>=7.0,<8.0

# Utilities
tqdm>=4.66,<5.0

# Dimensionality reduction
joblib>=1.3,<1.5
